First separated signature implementation
In that commit the previous signature logic in the completions provider
is disabled, instead a new module handles the edit and cursor moves and
shows the signature panel when it happens to be inside a function call,
this is just a way to quickly get a signature and experiment with the
completions panel at the same time.

The cohabitation with completion is done using a mutation observer. As
we can't get any information from the completion plugin, we listen to
the change of its `styles` attributes and check when it start
overlapping our signature to swap its place.